feat: Add `generate` and `streaming-input` feature toggles.
That way, it's possible to not compile a bunch of code in `gix` if
the writing of packs isn't required.
