+++ gix-pack/Cargo.toml
@@ -14,7 +14,11 @@ autotests = false
 doctest = false
 
 [features]
-
+default = ["generate", "streaming-input"]
+## generate new packs from a set of objects.
+generate = ["dep:gix-traverse", "dep:gix-diff"]
+## Receive a pack as datastream and resolve it
+streaming-input = []
 ## Provide a fixed-size allocation-free LRU cache for packs. It's useful if caching is desired while keeping the memory footprint
 ## for the LRU-cache itself low.
 pack-cache-lru-static = ["dep:uluru"]
@@ -25,7 +29,7 @@ object-cache-dynamic = ["dep:clru"]
 ## Data structures implement `serde::Serialize` and `serde::Deserialize`.
 serde = ["dep:serde", "gix-object/serde"]
 ## Make it possible to compile to the `wasm32-unknown-unknown` target.
-wasm = ["gix-diff/wasm"]
+wasm = ["gix-diff?/wasm"]
 
 [dependencies]
 gix-features = { version = "^0.33.0", path = "../gix-features", features = ["crc32", "rustsha1", "progress", "zlib"] }
@@ -33,14 +37,18 @@ gix-path = { version = "^0.9.0", path = "../gix-path" }
 gix-hash = { version = "^0.12.0", path = "../gix-hash" }
 gix-chunk = { version = "^0.4.4", path = "../gix-chunk" }
 gix-object = { version = "^0.35.0", path = "../gix-object" }
-gix-traverse = { version = "^0.31.0", path = "../gix-traverse" }
-gix-diff = { version = "^0.34.0", path = "../gix-diff" }
 gix-hashtable = { version = "^0.3.0", path = "../gix-hashtable" }
+
+# for streaming of packs (input, output)
+gix-traverse = { version = "^0.31.0", path = "../gix-traverse", optional = true }
+gix-diff = { version = "^0.34.0", path = "../gix-diff", default-features = false, optional = true }
 
 memmap2 = "0.7.0"
 smallvec = "1.3.0"
 parking_lot = { version = "0.12.0", default-features = false }
 thiserror = "1.0.26"
+
+# for caching
 uluru = { version = "3.0.0", optional = true }
 clru = { version = "0.6.1", optional = true }
