+++ gix-pack/src/data/input/entries_to_bytes.rs
@@ -73,12 +73,11 @@ where
         }
         self.num_entries += 1;
         entry.header.write_to(entry.decompressed_size, &mut self.output)?;
-        std::io::copy(
-            &mut entry
+        self.output.write_all(
+            entry
                 .compressed
                 .as_deref()
                 .expect("caller must configure generator to keep compressed bytes"),
-            &mut self.output,
         )?;
         Ok(entry)
     }
