+++ gix-pack/src/data/mod.rs
@@ -37,9 +37,11 @@ pub mod init {
 pub mod entry;
 
 ///
+#[cfg(feature = "streaming-input")]
 pub mod input;
 
 /// Utilities to encode pack data entries and write them to a `Write` implementation to resemble a pack data file.
+#[cfg(feature = "generate")]
 pub mod output;
 
 /// A slice into a pack file denoting a pack entry.
