+++ gix-pack/src/index/util.rs
@@ -1,4 +1,4 @@
-use std::{io, time::Instant};
+use std::time::Instant;
 
 use gix_features::progress::{self, Progress};
 
@@ -18,30 +18,4 @@ pub(crate) fn index_entries_sorted_by_offset_ascending(
 
     progress.show_throughput(start);
     v
-}
-
-pub(crate) struct Count<W> {
-    pub bytes: u64,
-    pub inner: W,
-}
-
-impl<W> Count<W> {
-    pub fn new(inner: W) -> Self {
-        Count { bytes: 0, inner }
-    }
-}
-
-impl<W> io::Write for Count<W>
-where
-    W: io::Write,
-{
-    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-        let written = self.inner.write(buf)?;
-        self.bytes += written as u64;
-        Ok(written)
-    }
-
-    fn flush(&mut self) -> io::Result<()> {
-        self.inner.flush()
-    }
 }
